Strict checking before adding a FlowPropertyValueProvider to a FlowPropertyDefinition
when baseFlowManagement needs to create a temp property it goes through FlowDefinitionsManager.getFlowPropertyDefinitionBuilder()
this enables standard {propertyName,class} combinations to be registered.

FlowPropertyDefinition now strictly checks to make sure that a change is allowed. and the various init*() now return clones if the FPD is a templateFlowPropertyDefinition

And lots more tests!